requirements.txt
Flask==3.0.0
PyJWT==2.8.0
ConfigParser==6.0.0